backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.4
python-dotenv==1.1.0
langgraph==0.4.1
langchain-core==0.3.56
langchain-groq==0.3.2
SQLAlchemy==2.0.40
